+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Support 2023.2 IDEs
 - Fix leo statement recover rule
 - Fix exceptions on code analysis
+- Add type checking for method calls and associated functions
 
 ## [0.7.0]